More misc. cleanup
- Import `css` from DOM module instead of functions
- Add new popup module
- Move `between()` to math module
